Added logic to detect directories with content
Instead of overwriting batch files, it will now check whether a directory is empty first and abort if it is not empty. A new command line flag -f/--force-overwrite was added to ignore existing ".sh" files. It will still abort if any other kind of file exists in the folder.